First version of the Snice Codecs Base
Restructuring and moving code from snice-networking into this base codec
library so that e.g. pkts.io has an easier time to consume it.

Also, some codecs, such as Diameter & GTP, have a lot of the classes
auto generated and it is somewhat of a catch-22 where we do want to have
a set of handwritten base classes that the code generation will use but
at the same time, the main codec library needs to that dependency as
well and so does the codegen library and as such, they can't really be
in the same repo.

Hence, for other codecs such as SIP, it won't necessarily have anything
in the base codec library.