Added a request/response to the type.
A new attribute on the MessageTypeMetaData for indicating if the
mesesage is a request or response. This is being used by the code
generators. Will eventually use it to drive the createResponse on a
request but will then also need to know which response to create since
there simply isn't a request/response bit to flip in the header. Oh GTP,
you silly protocol...