Decided to include the wireshark diamter xml files within the jar.
Turns out that asking other to checkout wireshark to get this to build,
or rather the snice-codecs to use the diameter-codegen plugin, which
would then refer to the wireshark directory to find these files, is a
bit cumbersom and annoying. Therefore, opted to include those xml files
within the jar itself so everything is self contained.